+++ common/src/main/java/com/majruszsaccessories/accessories/AnglerRune.java
@@ -12,7 +12,7 @@ public class AnglerRune extends AccessoryHandler {
 	public AnglerRune() {
 		super( MajruszsAccessories.ANGLER_RUNE, AnglerRune.class );
 
-		this.add( FishingLuckBonus.create( 3 ) )
+		this.add( FishingLuckBonus.create( 3.5f ) )
 			.add( FishingLureBonus.create( 0.25f ) )
 			.add( FishingExtraItems.create( 0.25f, 3 ) );
 	}