+++ common/src/main/java/com/majruszsaccessories/accessories/AnglerTrophy.java
@@ -24,7 +24,7 @@ public class AnglerTrophy extends AccessoryHandler {
 	public AnglerTrophy() {
 		super( MajruszsAccessories.ANGLER_TROPHY, AnglerTrophy.class );
 
-		this.add( FishingLuckBonus.create( 3 ) )
+		this.add( FishingLuckBonus.create( 3.0f ) )
 			.add( FishingDropChance.create( 0.01f ) )
 			.add( TradeOffer.create() );
 	}