+++ changelog.md
@@ -1,4 +1,6 @@
 - added extra Pocket slot for Trinkets and Curios API
 - added support for multiple accessories
+- changed fishing luck bonus from an integer to a floating point number
 - removed unused code
+- increased Angler Rune luck bonus from 3 to 3.5
 - fixed bug with Horseshoe luck bonus being affected by efficiency bonus